requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
web3==6.15.1
eth-abi==5.0.1
eth-account==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
alembic==1.12.1